Parse CSV files into typed data
Use the `csv` library to parse the CSV files into arrays of JS objects
where the column names are the keys of the object and rows of the CSV
are the values. Then, use `zod` to define types for these parsed logs
and to validate their schema.

In doing this I discovered a bug in the logs exported by STFC which I've
had to work around in order to successfully parse them. I've tried
raising it in their discord but I'm not sure if that has stuck. I'll
probably try to submit it via the chat bot eventually.