Fix file split error
The file splitting logic was broken, and stripped off the first line of
all the output split log files after the first. This fixes that and adds
tests for it.

Added a couple more examples of combat logs and tests to make sure the
correct number of output files are produced.

I only thoroughly test one of the combat logs to save time.