Refactor log parser
Hide the implementation details from the armada router and enable better
organized tests.

This ended up being a much bigger changeset than I anticipated because
this refactor found a bug in the parsing logic where if one of the CSVs
had an empty data set (e.g. the combat was a defeat so we gained no
loot/rewards) the schema was not detected properly. So now we do some
special handling to first detect the schema, then we parse the file
with the correct parser, keeping track of the type of the data even if
it's empty.